test: 공통된 테스트를 @ParameterizedTest로 묶음
