+++ src/test/java/webserver/message/MediaTypeTest.java
@@ -1,8 +1,7 @@
-package webserver.message.request;
+package webserver.message;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.message.MediaType;
 
 import static org.assertj.core.api.Assertions.assertThat;
